requirements.txt
oandapyV20
python-dotenv
pandas
numpy
scikit-learn
matplotlib
